requirements.txt
numpy
pandas
torch
torchvision
mlflow
scikit-learn
tensorflow
tensorboard
Flask
Pillow
